Move Slackapp module to its own folder
